+++ NEWS.md
@@ -1,5 +1,8 @@
 # News
 
+## Unreleased
+  * Added: BREAKING CHANGE: Reject `sequence` definitions for Active Record primary keys
+
 ## 6.2.1 (March 8, 2022)
   * Added: CI testing against truffleruby
   * Changed: Documentation improvements for sequences and traits